moved commands to their own files
