+++ Bess/src/scene/scene.cpp
@@ -11,7 +11,8 @@
 #include "gtc/type_ptr.hpp"
 #include "pages/main_page/main_page_state.h"
 #include "scene/artist.h"
-#include "scene/commands/commands.h"
+#include "scene/commands/connect_command.h"
+#include "scene/commands/delete_comp_command.h"
 #include "scene/components/components.h"
 #include "scene/renderer/renderer.h"
 #include "settings/viewport_theme.h"
